:hammer: refactor: 일기 리뷰 API ImageController에 이미지 리뷰 API로 변경
