+++ src/main/java/tipitapi/drawmytoday/domain/diary/controller/ImageController.java
@@ -8,14 +8,18 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import tipitapi.drawmytoday.common.resolver.AuthUser;
 import tipitapi.drawmytoday.common.security.jwt.JwtTokenInfo;
+import tipitapi.drawmytoday.domain.diary.dto.ReviewDiaryRequest;
 import tipitapi.drawmytoday.domain.diary.service.ImageService;
 
 @RestController
@@ -58,6 +62,30 @@ public class ImageController {
         @Parameter(description = "일기 이미지 id", in = ParameterIn.PATH) @PathVariable("id") Long imageId
     ) {
         imageService.deleteImage(tokenInfo.getUserId(), imageId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @Operation(summary = "일기 평가", description = "주어진 ID의 일기를 평가한다.")
+    @ApiResponses(value = {
+        @ApiResponse(
+            responseCode = "204",
+            description = "성공적으로 일기를 평가함"),
+        @ApiResponse(
+            responseCode = "403",
+            description = "I004 : 자신의 이미지에만 접근할 수 있습니다.",
+            content = @Content(schema = @Schema(hidden = true))),
+        @ApiResponse(
+            responseCode = "404",
+            description = "I001 : 이미지를 찾을 수 없습니다.",
+            content = @Content(schema = @Schema(hidden = true))),
+    })
+    @PostMapping("/{id}/review")
+    public ResponseEntity<Void> reviewImage(
+        @AuthUser JwtTokenInfo tokenInfo,
+        @Parameter(description = "이미지 id", in = ParameterIn.PATH) @PathVariable("id") Long imageId,
+        @RequestBody @Valid ReviewDiaryRequest reviewDiaryRequest
+    ) {
+        imageService.reviewDiary(tokenInfo.getUserId(), imageId, reviewDiaryRequest.getReview());
         return ResponseEntity.noContent().build();
     }
 }
